backend/pyproject.toml
[project]
name = "doc-assistant"
version = "0.1.0"
description = "公文助手 — 本地化政务公文知识库与 AI 写作辅助系统"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.0",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "anthropic>=0.30",
    "chromadb>=0.5",
    "python-docx>=1.0",
    "pdfplumber>=0.11",
    "pytesseract>=0.3",
    "openpyxl>=3.1",
    "python-pptx>=0.6",
    "py7zr>=0.22",
    "pyunpack>=0.3",
    "Pillow>=10.0",
    "pdf2image>=1.16",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
